Changes for Publication standard v0.4.0

+++ backend/app/etl/resources/loader.py
@@ -3,50 +3,24 @@ import numpy as np
 import json
 from pydantic import ValidationError
 from app.database.database import SessionLocal
-from app.etl.config.excel_column_mapping import excel_column_mapping
 from app import models, schemas
-from app.util import upc
+from app.util.logger import get_logger
+from app.config.settings import Settings
+
+logger = get_logger(__name__)
+env_settings = Settings()
 
 
 class AlgoritmeLoader:
     """Load algoritmes from a json file. Existing algoritmes will be removed."""
 
-    def __init__(self, excel_file: str | None = None, json_file: str | None = None):
-        if excel_file is not None:
-            df = self.__get_df_algoritme_from_excel(excel_file)
-        elif json_file is not None:
+    def __init__(self, json_file: str | None = None):
+        if json_file is not None:
             df = self.__get_df_algoritme_from_json(json_file=json_file)
         else:
             raise RuntimeError("Either excel_file or json_file must be specified")
 
         self.__algoritmes = self.process_df(df)
-
-    @staticmethod
-    def __get_df_algoritme_from_excel(excel_file: str) -> pd.DataFrame:
-        sheet_names = pd.read_excel(excel_file, sheet_name=None).keys()
-        included_sheet_names = [
-            sheet_name
-            for sheet_name in sheet_names
-            if sheet_name not in ["Template (dupliceer dit blad)", "Lege invullijst"]
-        ]
-        sheet_name = included_sheet_names[0]
-        df_all = pd.DataFrame()
-        for sheet_name in included_sheet_names:
-            df = (
-                pd.read_excel(excel_file, sheet_name=sheet_name, header=None)
-                .drop(columns=[0])
-                .set_index(1)
-                .T
-            )
-            df_all = pd.concat([df_all, df])
-
-        df_all["lars"] = None
-        df_all["published"] = True
-        df_all["released"] = True
-        df_all["Schema"] = "0.1.0"
-        df_all["Herzieningsdatum"] = df_all["Herzieningsdatum"].astype("string")
-
-        return df_all.rename(columns=excel_column_mapping)
 
     @staticmethod
     def __get_df_algoritme_from_json(json_file: str) -> pd.DataFrame:
@@ -91,64 +65,47 @@ class AlgoritmeLoader:
     def load_algoritmes(self):
         counter = 0
         with SessionLocal() as db:
+            # Query the AlgoritmeVersion table, which is only populated at the end of
+            logging_algos = db.query(models.AlgoritmeVersion).all()
+            if (len(logging_algos) == 0) and (env_settings.type != "DEV"):
+                logger.info(
+                    "No algorithms in database. It is assumed another pod is running the ETL. Stopping loading."
+                )
+                return False
             db.query(models.Algoritme).delete()
             db.commit()
-            for new_version in self.__algoritmes:
-                # Checks presence in db based on LARS-code.
-                lars = new_version["lars"]
-                if lars:
-                    algo_in_db = (
-                        db.query(models.Algoritme)
-                        .filter(models.Algoritme.lars == lars)
-                        .first()
-                    )
-                else:
-                    algo_in_db = None
+            for algo in self.__algoritmes:
+                lars = algo["lars"]
 
-                if not algo_in_db:
-                    # If not found, makes a new algorithm.
+                algoritme_version_dict = algo.copy()
+                algoritme_version_dict.pop("lars")
+                algoritme_version_dict.pop("owner")
+                new_algoritme_version = models.AlgoritmeVersion(
+                    **algoritme_version_dict
+                )
 
-                    # Test provided LARS-code for validity.
-                    lars_is_valid = upc.validate_upc(new_version["lars"])
-                    if not lars_is_valid:
-                        # Redefines LARS-code.
-                        lars_list = [
-                            id[0] for id in db.query(models.Algoritme.id).all()
-                        ]
-                        lars = upc.find_new_upc(avoid_upc_list=lars_list)
+                # Makes a new algorithm.
+                algoritme_dict = {
+                    "lars": lars,
+                    "owner": algo["owner"],
+                    "versions": [new_algoritme_version],
+                }
+                new_algoritme = models.Algoritme(**algoritme_dict)
 
-                    # Makes a new algorithm.
-                    algoritme_dict = {
-                        "lars": lars,
-                        "owner": new_version["owner"],
-                    }
-                    new_algoritme = models.Algoritme(**algoritme_dict)
-                    db.add(new_algoritme)
-                    db.flush()
-                    db.refresh(new_algoritme)
-                    algoritme_id = new_algoritme.id
-                else:
-                    # If found, archive newest algo
-                    algoritme_id = algo_in_db.id
+                db.add(new_algoritme)
+                db.add(new_algoritme_version)
+                logger.info(
+                    "Added " + str(lars) + " | " + str(new_algoritme_version.name)
+                )
 
-                # add a new algorithm_version
-                # Remove all the columns that belong in the 'algoritme' table.
-                new_version.pop("lars")
-                new_version.pop("owner")
-                new_version["algoritme_id"] = algoritme_id
-                new_version_model = models.AlgoritmeVersion(**new_version)
-
-                # test validity
-
-                version_str = "v" + new_version["standard_version"].replace(".", "_")
+                # Tests validity.
+                version_str = "v" + algo["standard_version"].replace(".", "_")
                 schema = schemas.versions.create_algorithm_in_loader_schema(version_str)
                 try:
-                    schema.parse_obj(new_version)
+                    schema.parse_obj(algo)
                 except ValidationError as e:
-                    print(e)
+                    logger.warning(e)
                     counter += len(e.args[0])
-
-                db.add(new_version_model)
-                db.commit()
-        print("Number of validation errors:", counter)
+            db.commit()
+        logger.info("Number of validation errors: " + str(counter))
         return True
